Added ExportData.Record(). Removed OnCurrentWorkspaceChanged.
Moved collection of environment data to  new ExportData.Record() method.
Removed OnCurrentWorkspaceChanged events.
