+++ BinocularsViewExtension.cs
@@ -20,7 +20,6 @@ namespace Binoculars
         public string UniqueId => "4DB6C8D9-7D8E-42A8-8995-E14ACFA037CF";
         public string Name => "Binoculars View Extension";
 
-
         private string UserName = Environment.UserName;
 
         private MenuItem _extensionMenu;
@@ -44,6 +43,16 @@ namespace Binoculars
         /// </param>
         public void Loaded(ViewLoadedParams vlp)
         {
+
+            // Display a MessageBox to the user
+            // @todo Use a view framework to improve the UI/UX
+            string message = "By pressing OK you agreeing to Binoculars 🔍 data collection.";
+            string title = "Terms of Use Agreement";
+            MessageBox.Show(message, title, MessageBoxButton.OK, MessageBoxImage.Exclamation);
+
+            // Collect the environment variables
+            ExportData.Collect(vlp);
+
             // hold a reference to the Dynamo params to be used later
             _viewLoadedParams = vlp;
             
@@ -56,63 +65,8 @@ namespace Binoculars
         }
 
         /// <summary>
-        /// Method that is called when Dynamo has finished loading and is ready to be used.
-        /// </summary>
-        /// <param name="rp">
-        /// Parameters that provide references to Dynamo commands, settings and events.
-        /// This object is supplied by Dynamo itself.
-        /// </param>
-        public void Ready(ReadyParams rp)
-        {
-            // @todo Subscribe to an event after the Dynamo home screen is shown
-
-            // Display a MessageBox to the user
-            // @todo Use a view framework to improve the UI/UX
-            string message = "By pressing OK you agreeing to Binoculars 🔍 data collection.";
-            string title = "Terms of Use Agreement";
-            MessageBox.Show(message, title, MessageBoxButton.OK, MessageBoxImage.Exclamation);
-
-            // @todo if user does not consent, don't store..
-            if (true)
-            {
-
-                // Dynamo StartupParams
-                ExportData.dynamoVersion = rp.StartupParams.DynamoVersion.ToString();
-
-                // @todo Determine whether the script is run from DynamoPlayer..
-
-                // Store the local environment data
-                ExportData.user = Environment.MachineName;
-                ExportData.computerName = Environment.UserName;
-
-                // @todo Define the revitVersion, leave blank or null if opened from any other environment (Sandbox, Civil3D etc)
-                ExportData.revitVersion = "2019.0.2";
-
-                // Request the IP and geolocation from ipinfo.io
-                WebClient client = new WebClient();
-                client.Headers.Set("Accept", "application/json");
-                var json = client.DownloadString("https://ipinfo.io");
-                // @todo We need to check the request was actually sent successfully and gracefully deal with cases where the API is unavailable
-
-                // Parse the JSON
-                JObject ipinfo = JObject.Parse(json);
-
-                // Store geolocation data
-                ExportData.ip = (string)ipinfo["ip"];
-                ExportData.latlng = (string)ipinfo["loc"];
-                ExportData.city = (string)ipinfo["city"];
-                ExportData.country = (string)ipinfo["country"];
-            }
-
-            // we can register our own events that will be triggered when specific things happen in Dynamo
-            // a reference to the ReadyParams is needed to do this, so we pass it on
-            BinocularsEvents.RegisterEventHandlers(rp);
-        }
-
-        /// <summary>
         /// Adds custom menu items to the Dynamo menu
         /// </summary>
-
         public string ToTitleCase(string str)
         {
             return CultureInfo.CurrentCulture.TextInfo.ToTitleCase(str.ToLower());
@@ -151,7 +105,6 @@ namespace Binoculars
         /// </summary>
         public void Shutdown()
         {
-
             BinocularsEvents.UnregisterEventHandlers();
         }
